updates
-run as a script, not from cmd line
-loops for all files in the current directory
-only processes image files
-writes to a text file named "results.txt", appending each time. File is created if it does not already exist in the current directory
-no longer opens image files at the end of the script
-no longer saves any temporary working images files (grayscale, png)

+++ image_text_extraction_tesseract.py
@@ -1,42 +1,29 @@
 import cv2
-import os,argparse
+import os
 import pytesseract
 from PIL import Image
  
-#pytesseract.pytesseract.tesseract_cmd = r'C:\Program Files\Tesseract-OCR'
-pytesseract.pytesseract.tesseract_cmd = 'C:\\Program Files\\Tesseract-OCR\\tesseract.exe'
+pytesseract.pytesseract.tesseract_cmd = "C:\\Program Files\\Tesseract-OCR\\tesseract.exe"
+current_path = os.listdir(os.getcwd())
+border = "********************"
+ 
+for images in current_path:
+    #convert to grayscale image
+    img = cv2.imread(images)
+    if img is None:
+        continue
 
+    gray=cv2.cvtColor(img, cv2.COLOR_BGR2GRAY)
+    gray=cv2.threshold(gray, 0,255,cv2.THRESH_BINARY| cv2.THRESH_OTSU)[1]
+    gray=cv2.medianBlur(gray, 3)
+        
+    #memory usage with image i.e. adding image to memory
+    filename = "{}".format(str(images))
+    cv2.imwrite(filename, gray)
+    text = pytesseract.image_to_string(Image.open(filename))
+    os.remove(filename)
+    #print(text)
+    with open("results.txt", "a+") as file:
+        file.write("%s\n%s\n%s\n%s\n\n" % (border, filename, border, text))
 
-#We then Construct an Argument Parser
-ap=argparse.ArgumentParser()
-ap.add_argument("-i","--image",
-                required=True,
-                help="Path to the image folder")
-ap.add_argument("-p","--pre_processor",
-                default="thresh", 
-                help="the preprocessor usage")
-args=vars(ap.parse_args())
- 
-#We then read the image with text
-images=cv2.imread(args["image"])
- 
-#convert to grayscale image
-gray=cv2.cvtColor(images, cv2.COLOR_BGR2GRAY)
- 
-#checking whether thresh or blur
-if args["pre_processor"]=="thresh":
-    cv2.threshold(gray, 0,255,cv2.THRESH_BINARY| cv2.THRESH_OTSU)[1]
-if args["pre_processor"]=="blur":
-    cv2.medianBlur(gray, 3)
-     
-#memory usage with image i.e. adding image to memory
-filename = "{}.jpg".format(os.getpid())
-cv2.imwrite(filename, gray)
-text = pytesseract.image_to_string(Image.open(filename))
-os.remove(filename)
-print(text)
- 
-# show the output images
-cv2.imshow("Image Input", images)
-cv2.imshow("Output In Grayscale", gray)
-cv2.waitKey(0)
+file.close()
